==> experiments_506/0.clean_issues_&_EDA.ipynb <==
# Clean Issues Query Dataset

The dataset "042925_issues_query.csv" is provided by clearinghouse through their internal SQL query. This dataset contains the case issue and issue category for each case. This notebook outlines the steps I undertook to clean this dataset for building the ML model to predict case issues/issue categories.

# Import Libraries

import json
import pandas as pd

# Load the dataset

issues = pd.read_csv("data/042925_issues_query.csv")
issues.head()

# Clean the dataset

## The case_issue_info is in json format, parse it so we can separate out the issue category, case type, and description

def parse_json_column(val):
    if pd.isna(val) or val == '':
        return {}
    try:
        return json.loads(val)
    except (ValueError, SyntaxError):
        return {}

parsed = issues['case_issue_info'].apply(parse_json_column)
expanded = pd.json_normalize(parsed)
issues = issues.join(expanded[['category', 'case_type', 'description']])

issues.head()

assert issues["case_issue_info"].isna().sum() == issues["category"].isna().sum()

## We do not care about case type and description, discard these along with the case_issue_info column as we already parsed it

issues = issues[["case_id", "doc_id", "case_issue", "category"]]

## Get a better understanding of the relationship between doc_id and case_issue

issues[issues["case_id"] == 45726]

issues[issues["case_id"] == 17099]

I sampled some docs from the sampled cases to review them and don't believe there is any relationship between the doc_id and the case_issue. The case_issue is attached to the case level, not specifically associated with a single document. The apparent one-to-one mapping between doc_id and case_issue is an artificat from the join. 

## Remove the doc_id since we want to focus on the issues at case level.

issues = issues[["case_id", "case_issue", "category"]]

## Rename the category column to issue_category to avoid confusion

issues.rename(columns={'category': 'issue_category'}, inplace=True)

## Drop any rows that don't contain case_issue or issue_category & remove duplicates

assert issues["case_issue"].isna().sum() == issues["issue_category"].isna().sum()
len(issues)

issues = issues[~issues["case_issue"].isna()].drop_duplicates()
len(issues)

issues.head()

## Do some preliminary EDA

- ~12K unique cases
- 405 unique issues, belonging to 22 issue categories
- Some cases have as many as 42 issues, but most cases have fewer than 10 issues
- Some cases have as many as 9 issue categories, but most cases have fewer than 5 issue categories

issues["case_id"].nunique()

issues["case_issue"].nunique()

issues["issue_category"].nunique()

issues[["case_id", "case_issue"]].groupby("case_id").nunique().describe()

issues[["case_id", "issue_category"]].groupby("case_id").nunique().describe()

## Most issues belong to only one category, with the exception of 4 issues, obtain an understanding of why

issue_category_counts = issues.groupby('case_issue')['issue_category'].nunique()
issue_category_counts[issue_category_counts > 1]

## Issue == Funding

issues[issues["case_issue"] == "Funding"]["issue_category"].value_counts()

Ideally, within one case, we should have a one-to-one match between issue and issue category, find out any cases that have multiple matches

multi_match = issues[issues["case_issue"] == "Funding"][["case_id","issue_category"]].groupby("case_id").nunique()
multi_match[multi_match["issue_category"] > 1]

issues[(issues["case_issue"] == "Funding") & (issues["case_id"] == 45997)]

issues[issues["case_id"] == 45997]["case_issue"].value_counts()

issues[issues["case_id"] == 45997]["issue_category"].value_counts()

Case id 45997 appears to be mostly related to "Reproductive rights", for consistency, we'll drop "General/Misc." issue category.

issues.drop(10554, inplace=True)

multi_match = issues[issues["case_issue"] == "Funding"][["case_id","issue_category"]].groupby("case_id").nunique()
assert len(multi_match[multi_match["issue_category"] > 1]) == 0

## Issue == Spanish

issues[issues["case_issue"] == "Spanish"]["issue_category"].value_counts()

Ideally, within one case, we should have a one-to-one match between issue and issue category, find out any cases that have multiple matches

multi_match = issues[issues["case_issue"] == "Spanish"][["case_id","issue_category"]].groupby("case_id").nunique()
multi_match[multi_match["issue_category"] > 1]

## Issue == Testing

issues[issues["case_issue"] == "Testing"]["issue_category"].value_counts()

Ideally, within one case, we should have a one-to-one match between issue and issue category, find out any cases that have multiple matches

multi_match = issues[issues["case_issue"] == "Testing"][["case_id","issue_category"]].groupby("case_id").nunique()
multi_match[multi_match["issue_category"] > 1]

for case_id in multi_match[multi_match["issue_category"] > 1].index.to_list():
    display(issues[(issues["case_issue"] == "Testing") & (issues["case_id"] == case_id)])

issues[issues["case_id"] == 829]["case_issue"].value_counts()

issues[issues["case_id"] == 829]["issue_category"].value_counts()

All the cases appear to be assigned to both "Discrimination Area" and "General/Misc.
", for consistency, we'll drop "General/Misc." issue category.

issues = issues[~((issues['case_issue'] == 'Testing') & (issues['issue_category'] == 'General/Misc.'))]

multi_match = issues[issues["case_issue"] == "Testing"][["case_id","issue_category"]].groupby("case_id").nunique()
assert len(multi_match[multi_match["issue_category"] > 1]) == 0

## Issue == Other

issues[issues["case_issue"] == "Other"]["issue_category"].value_counts()

Ideally, within one case, we should have a one-to-one match between issue and issue category, find out any cases that have multiple matches

multi_match = issues[issues["case_issue"] == "Other"][["case_id","issue_category"]].groupby("case_id").nunique()
multi_match[multi_match["issue_category"] > 1]

for case_id in multi_match[multi_match["issue_category"] > 1].index.to_list():
    display(issues[(issues["case_issue"] == "Other") & (issues["case_id"] == case_id)])

It's reasonable that multiple issue categories could have "other" issue, so leave these as is.

## Do another round of EDA

issue_category_counts = issues.groupby('case_issue')['issue_category'].nunique()
issue_category_counts[issue_category_counts > 1]

issues.groupby('issue_category')['case_issue'].nunique()

## For each case_id and each issue_category, produce the list of case_issue

grouped = issues.groupby(["case_id", "issue_category"])["case_issue"].agg(list).reset_index()
grouped.head()

## Confirm no duplicate issue_category for each case_id

case_category_count = grouped[["case_id", "issue_category"]].groupby("case_id").value_counts()
assert len(case_category_count[case_category_count > 1]) == 0

## To make it easier to map case to issue category and issues, put the issue data into a dict

issue_data = grouped.groupby('case_id').apply(
    lambda x: dict(zip(x['issue_category'], x['case_issue']))
).reset_index(name='issue_data')

issue_data.head()

assert len(issue_data) == issue_data["case_id"].nunique()

len(issue_data)

## Save the issue data for future use

issue_data.to_json("data/case_issues.json")
