==> experiments_506/1.clean_data_&_EDA.ipynb <==
# Clean Clearinghouse Dataset

The dataset "clearinghouse.csv" is provided by clearinghouse through their internal SQL query. This dataset contains the case metadata for each case that has a complete coding. This notebook outlines the steps I undertook to clean this dataset for building the ML model to predict case issues/issue categories.

# Import Libaries

import pandas as pd

from sklearn.preprocessing import MultiLabelBinarizer

# Join the issues dataset to the clearinghouse dataset that contains all the other metadatas

issue_data = pd.read_json("data/case_issues.json")

cases = pd.read_csv("data/clearinghouse.csv")
len(cases)

result = cases.merge(issue_data, on="case_id", how="left")
len(result)

result.head()

## Check for cases with black issue data

result[(result["issue_data"].isna()) & (~result["case_issues"].isna())]

result[(~result["issue_data"].isna()) & (result["case_issues"].isna())]

## Drop all cases where the issue_data is blank, as we cannot use them to make predictions / evaluations

len(result)

result = result[~result["issue_data"].isna()]
len(result)

## Extract the issue categories and issues from issue data

result['issue_category'] = result["issue_data"].apply(lambda d: list(d.keys()) if isinstance(d, dict) else [])
result['issues'] = result["issue_data"].apply(
    lambda d: sorted([item for sublist in d.values() for item in sublist]) if isinstance(d, dict) else []
)

result.columns

## Drop columns we don't need

result = result[['case_id', 'case_name', 'case_status', 'case_state', 'court_name',
       'case_type', 'case_ongoing', 'case_special_collections',
       'case_causes_of_action', 'issue_category', 'issues']]
result.head()

## Save the result df for future use

result.to_json("data/clean.json")

## Do some EDA on the result df

cols = ["case_state", "court_name", "case_type", "case_ongoing", "case_special_collections"]

for col in cols:
    print("---", col, "---")
    display(result[col].nunique())
    display(result[col].value_counts())

mlb = MultiLabelBinarizer()
binary_matrix = mlb.fit_transform(result['issue_category'].to_list())
binary_matrix.shape

category_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)
category_df.sum().sort_values(ascending=False)

mlb = MultiLabelBinarizer()
binary_matrix = mlb.fit_transform(result['issues'].to_list())
binary_matrix.shape

issues_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)
issues_df.sum().sort_values(ascending=False)

issues_counts = issues_df.sum()

threshold = issues_counts.quantile(0.90)
issues = issues_counts[issues_counts >= threshold]
issues.sort_values(ascending=False)
