==> experiments_506/2.extract_documents.ipynb <==
# Clean Document Dataset

The document dataset 042925_documents_query.csv was provided by clearinghouse through their internal SQL query. However, the file is too large to efficiently work with, so this notebook outlines the steps I undertook to make the file more workable.

# Import Libaries

import pandas as pd

# Take a peak at the data

df = pd.read_csv('data/042925_documents_query.csv', nrows=10)
df.head()

# Save the ocr text to plain text files & remove the ocr text and columns that we don't need

cols = ['case_id', 'doc_id', 'doc_title', 'doc_date', 'doc_type', 'doc_source', 'doc_url', 'doc_ocr_text']

with pd.read_csv('data/042925_documents_query.csv', chunksize=1000, usecols=cols) as reader:
    for i, chunk in enumerate(reader):
        # Save each doc_ocr_text to a separate file
        for _, row in chunk.iterrows():
            filename = f"docs/{row['case_id']}_{row['doc_id']}.txt"
            with open(filename, 'w', encoding='utf-8') as f:
                f.write(str(row['doc_ocr_text']))

        # Save the metadata to documents.csv
        chunk.loc[:, "doc_len"] = chunk["doc_ocr_text"].str.len()
        chunk = chunk[cols[:-1] + ['doc_len']]
        chunk.to_csv('data/documents.csv', mode='a', index=False, header=(i == 0))

# Load the document csv

docs = pd.read_csv("data/documents.csv")
docs.head()

# Clean the csv

## Remove all documents that don't have any content as they will not be useful for our task

len(docs)

docs = docs[~docs["doc_len"].isna()]
len(docs)

## Drop any duplicates

docs = docs.drop_duplicates()
len(docs)

## Add flags for identifing useful (and not useful) documents

docs.loc[:, "complaint_flag"] = docs['doc_title'].str.contains(r'\bcomplaint\b', case=False, na=False)
docs.loc[:, "opinion_flag"] = docs['doc_title'].str.contains(r'\bopinion\b', case=False, na=False)
docs.loc[:, "exhibit_flag"] = docs['doc_title'].str.contains(r'\bexhibit\b', case=False, na=False)

## Drop all exhibits as they do not contain useful information

docs = docs[~((docs["complaint_flag"] == 0) & (docs["opinion_flag"] == 0) & (docs["exhibit_flag"] == 1))]
docs[docs["exhibit_flag"] == 1]

## Save the csv for future use

docs.to_json("data/case_documents.json")
