==> experiments_506/3.join_data.ipynb <==
# Produce a dataset with both issues & documents, also filter for only those with either complaint or opinions

# Import Libaries

import pandas as pd

# Load the datasets

df = pd.read_json("data/clean.json")
docs = pd.read_json("data/case_documents.json")

df.head()

docs.head()

# Keep only cases that have both documents and assigned issues

cases = list(set(df["case_id"].unique()) & set(docs["case_id"].unique()))
len(cases)

len(df)

df = df[df["case_id"].isin(cases)]
len(df)

len(docs)

docs = docs[docs["case_id"].isin(cases)]
len(docs)

## Create a joined issue & document dataset for all future use

result = docs.merge(df, how="left", on="case_id")
len(result)

assert len(docs) == len(result)

result.head()

## Save the result df for future use

result.to_json("data/case_issues_documents.json")

## Do some EDA

result["case_id"].nunique()

Of ~10K cases, ~6.8K have at least one complaint document

result[result['complaint_flag'] == 1]['case_id'].nunique()

Of ~10K cases, ~8.3K have at least one complaint document or at least one opinion document

result[(result['complaint_flag'] == 1) | (result['opinion_flag'] == 1)]['case_id'].nunique()

## Given the complaint & opinion have the most useful information, filter for only complaints & opinions

meaningful = result[(result['complaint_flag'] == 1) | (result['opinion_flag'] == 1)]
len(meaningful)

meaningful.head()

Majority of the cases have at most 3 documents, given they could have more than one complaint or opinion for each case. Some cases have many documents, this is for cases that have numerous dockets that span across years.

meaningful["case_id"].value_counts().describe()

## Remove any documents have fewer than \~250 tokens (\~1000 chars) as they are unlikely to contain meaningful information for the prediction

meaningful = meaningful[meaningful["doc_len"] > (250 * 4)]
len(meaningful)

meaningful["case_id"].nunique()

## Save complaint & opinion df for future use

meaningful.to_json("data/case_with_complaint_opinion.json")
