==> experiments_506/4.clean_joined_data.ipynb <==
# Clean the joined data & look at the category distribution

# Import Libaries

import pandas as pd

from sklearn.preprocessing import MultiLabelBinarizer

# Load the dataset

df = pd.read_json("data/case_with_complaint_opinion.json")

## Do some EDA on the cases with complaint or opinion

subset = df[["case_id", "case_state", "court_name", "case_type", "case_ongoing", "issue_category", "issues"]]
subset["issue_category"] = subset["issue_category"].apply(tuple)
subset["issues"] = subset["issues"].apply(tuple)
subset = subset.drop_duplicates()
assert len(subset) == subset["case_id"].nunique()

cols = ["case_state", "court_name", "case_type", "case_ongoing"]

for col in cols:
    print("---", col, "---")
    display(subset[col].nunique())
    display(subset[col].value_counts())

## Look at the issue category distribution

mlb = MultiLabelBinarizer()
binary_matrix = mlb.fit_transform(subset['issue_category'].to_list())
binary_matrix.shape

category_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)
category_df.sum().sort_values(ascending=False)

## Given some categories have very few samples, select only the categories that have at least 200 cases so we have enough data for training

categories = category_df.sum().sort_values(ascending=False).index[:-5]
categories

len(categories)

len(df)

df['issue_category'] = df['issue_category'].apply(lambda labels: [label for label in labels if label in categories])
df = df[df['issue_category'].str.len() > 0].reset_index(drop=True)
len(df)

df["case_id"].nunique()

## Inspect the token distribution by case_id

Majority of the cases have more than 40K tokens, with very few cases have excessively long (and likely many) documents. Given the model context window 8,192 tokens, we need to apply some kind of aggregation on the embeddings by case such that we end up with one embedding per case.

token_df = df[["case_id", "doc_len"]].groupby("case_id").sum() / 4
token_df = token_df.rename(columns={"doc_len": "num_tokens"})

pd.set_option('display.float_format', '{:,.0f}'.format)  # No decimals, add commas
token_df.describe()

## Save the cleaned joined data for future use

df.to_json("data/clean_joined.json")
