==> experiments_506/5.split_data.ipynb <==
# Split the joined data to train/val/test sets for experiments

# Import Libaries

import random
random.seed(1234)

import pandas as pd

from sklearn.preprocessing import MultiLabelBinarizer

# Load the data

df = pd.read_json("data/clean_joined.json")
df.head()

## Given how imbalanced the categories are (some categories appear in many cases while other categories only appear in a handful of cases), the best way to produce a balanced dataset for training is to generate binary-classification dataset & models

cases = df[["case_id", "issue_category"]]
cases["issue_category"] = cases["issue_category"].apply(tuple)
cases = cases.drop_duplicates()
assert len(cases) == cases["case_id"].nunique()

mlb = MultiLabelBinarizer()
binary_matrix = mlb.fit_transform(cases['issue_category'].to_list())
binary_matrix.shape

category_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)
category_df.sum().sort_values(ascending=False)

category_df['case_id'] = cases['case_id'].values

## To ensure no case level data leakage, we will start by sampling from the least common categories, while keeping track of the case ids in train set

cats = category_df.sum().sort_values(ascending=False).index.drop("case_id")

for cat in cats[::-1]:
    print(cat)

train = {}
num_samples = 120

for cat in cats[::-1]:

    # Flatten all case_ids already in train
    train_ids = set(train_id for ids in train.values() for train_id in ids)
    
    # Get list of all case_ids for this category
    cat_ids = category_df[category_df[cat] == 1]['case_id'].to_list()

    # Filter cat_ids already in train_ids
    existing_ids = [cat_id for cat_id in cat_ids if cat_id in train_ids]

    # If we already have enough ids in train, do not add more, simply sample from the existing ids
    if len(existing_ids) >= num_samples:
        sampled_ids = random.sample(existing_ids, num_samples)
    # Otherwise, get more ids to be added to train_id
    else:
        remaining_ids = [cat_id for cat_id in cat_ids if cat_id not in train_ids]
        addl_ids = random.sample(remaining_ids, (num_samples - len(existing_ids)))
        sampled_ids = list(set(existing_ids + addl_ids))

    # Also get equal number of negative samples
    neg_ids = category_df[category_df[cat] == 0]['case_id'].to_list()
    neg_samples = random.sample(neg_ids, num_samples)
    
    train[cat] = list(set(sampled_ids + neg_samples))

for key, val in train.items():
    print(key)
    print(len(val))

train_ids = list(set(train_id for ids in train.values() for train_id in ids))
len(train_ids)

## We have sampled ~2K cases for training, now we will randomly sample ~800 cases from the rest of the cases to be put into val/test, while adhereing to the underlying distribution of categories & ensuring the cases in val are not in test

remaining_ids = category_df[~category_df['case_id'].isin(train_ids)]["case_id"].to_list()
len(remaining_ids)

val_ids = random.sample(remaining_ids, 800)
len(val_ids)

val_df = category_df[category_df['case_id'].isin(val_ids)]
val_df.sum().sort_values(ascending=False)

remaining_ids = category_df[~((category_df['case_id'].isin(train_ids)) | (category_df['case_id'].isin(val_ids)))]["case_id"].to_list()
len(remaining_ids)

test_ids = random.sample(remaining_ids, 800)
len(test_ids)

test_df = category_df[category_df['case_id'].isin(test_ids)]
test_df.sum().sort_values(ascending=False)

## Produce dataframes for train, val & test

train.keys()

abbreviations = {
    'COVID-19': 'COVID',
    'Benefits (Source)': 'BENEFITS',
    'LGBTQ+': 'LGBTQ',
    'Reproductive rights': 'REPRO',
    'Policing': 'POLICING',
    'Affected National Origin/Ethnicity(s)': 'NATION_ORIG',
    'Voting': 'VOTE',
    'Immigration/Border': 'IMMIGRATION',
    'Medical/Mental Health Care': 'MED',
    'Disability and Disability Rights': 'DISABILITY',
    'Affected Race(s)': 'RACE',
    'EEOC-centric': 'EEOC',
    'Jails, Prisons, Detention Centers, and Other Institutions': 'PRISON',
    'Affected Sex/Gender(s)': 'GENDER',
    'Discrimination Area': 'DISC_AREA',
    'Discrimination Basis': 'DISC_BASE',
    'General/Misc.': 'GENERAL'
}

filenames = []

for key, value in train.items():
    sub_df = df[df["case_id"].isin(value)]
    filename = f"data/train/{abbreviations[key]}.json"
    sub_df.to_json(filename)
    filenames.append(filename)

for i, filename in enumerate(filenames):
    reload = pd.read_json(filename)
    print("------")
    print(filename)
    print("number of records: ", len(reload))
    
    cases = reload[["case_id", "issue_category"]]
    cases["issue_category"] = cases["issue_category"].apply(tuple)
    cases = cases.drop_duplicates()
    
    mlb = MultiLabelBinarizer()
    binary_matrix = mlb.fit_transform(cases['issue_category'].to_list())
    category_df = pd.DataFrame(binary_matrix, columns=mlb.classes_)

    print("number of pos cases: ", len(category_df[category_df[list(train.keys())[i]] == 1]))
    print("number of neg cases: ", len(category_df[category_df[list(train.keys())[i]] == 0]))

sub_df = df[df["case_id"].isin(val_ids)]
sub_df.to_json("data/val/val.json")
print(len(sub_df))

sub_df = df[df["case_id"].isin(test_ids)]
sub_df.to_json("data/test/test.json")
print(len(sub_df))
