==> experiments_506/8a.covid_embeddings.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%pip install nupunkt -q

import ast
from tqdm import tqdm

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from nupunkt import sent_tokenize

import torch
from transformers import AutoTokenizer, AutoModel

MODEL_NAME = "answerdotai/ModernBERT-base"

TOKENIZER = AutoTokenizer.from_pretrained(MODEL_NAME)
MODEL = AutoModel.from_pretrained(MODEL_NAME)

df = pd.read_json('/content/drive/MyDrive/metadata/COVID.json')

df.head()

len(df)

MODEL.to('cuda')

BATCH_SIZE = 12

for index, row in tqdm(df.iterrows(), total=len(df)):
    case_id = row["case_id"]
    #print(f"Processing: ", index)

    chunks = ast.literal_eval(row["chunks"])

    all_embeddings = []

    for i in range(0, len(chunks), BATCH_SIZE):
        batch = chunks[i:i + BATCH_SIZE]

        # Tokenize and move inputs to GPU
        inputs = TOKENIZER(batch, max_length=8192, padding=True, truncation=True, return_tensors='pt')
        inputs = {k: v.to('cuda') for k, v in inputs.items()}

        # Run model on batch
        with torch.no_grad():
            outputs = MODEL(**inputs)
            last_hidden_state = outputs.last_hidden_state  # shape: [batch, seq_len, hidden_dim]
            #print("lhs: ", last_hidden_state.shape)

            mean_embeddings = last_hidden_state.mean(dim=1)  # shape: [batch, hidden_dim]
            all_embeddings.append(mean_embeddings.cpu())

        torch.cuda.empty_cache()

    # Combine all embeddings: [n_chunks, hidden_dim]
    all_embeddings_tensor = torch.cat(all_embeddings, dim=0)
    final_embedding = all_embeddings_tensor.mean(dim=0).numpy()
    #print("embed: ", final_embedding.shape)

    df.at[index, 'mean_embedding'] = final_embedding.astype(object)

    torch.cuda.empty_cache()

#df["max_embedding"] = df["max_embedding"].apply(lambda x: x.tolist() if isinstance(x, np.ndarray) else x)
df["mean_embedding"] = df["mean_embedding"].apply(lambda x: x.tolist() if isinstance(x, np.ndarray) else x)

df.to_json('/content/drive/MyDrive/metadata/embeddings/COVID_embeddings.json')

df.head()
