==> experiments_506/9.train_models.ipynb <==
# Train Models

Each notebook in the 9-series outlines the steps I undertook to train each binary classification model using the case content embeddings. The logistic regression models were trained through manual tuning by observing the training curves, the other models were trained through grid search to find the best hyperparameters based on validation performance.

All the trained models are saved in the models folder for future inference. Different issue categories may use different algorithms, depending on the validation performance.
